+++ package.json
@@ -21,6 +21,7 @@
     "webpack-dev-server": "^4.7.4"
   },
   "dependencies": {
+    "babel-eslint": "^10.1.0",
     "css-loader": "^6.6.0",
     "html-webpack-plugin": "^5.5.0",
     "lodash": "^4.17.21",
